+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 ## Unreleased
-- Add variables functionality.
+- Added variables functionality.
+- Console application now exits with the correct exit code.
+- Added line breaks every 70 characters in console applications progress bar.
 
 ## [1.2.0] - 2022-01-14
 - Changed required PHP version to >= 7.3.
 - Changed required datatypes version to 2.4.
 
 ## [1.1.0] - 2019-05-08
-- Add assert-header.
+- Added assert-header.
 
 ## 1.0.0 - 2019-04-03
 - First stable revision.